Refactor file fetchers to use Python's SharedFileFetcher
